+++ R/buoy.R
@@ -1,13 +1,44 @@
 #' Download buoy data
 #'
 #' This function downloads buoy data from various programs
-#' including...
+#' including, as listed in \sQuote{Details}.
+#'
+#' If `program` is `"MEDS"', then the choices for ID are as follows.
+#'
+#' * `"Banquereau Bank"`
+#' * `"East Scotian Slope"`
+#' * `"Halifax"`
+#' * `"Halifax DISCUS TriAx"`
+#' * `"Halifax Harbour"`
+#' * `"Laurentian Fan"`
+#' * `"Minas Basin"`
+#' * `"Port Hope"`
+#' * `"Prince Edward Point"`
+#' * `"Tail of the Bank"`
+#'
+#' If `program` is `"smartatlantic"', then the choices, all for
+#' buoys in the Atlantic Provinces, are as follows.
+#'
+#' * `"h1"` for a buoy near Duncan Reef, Nova Scotia
+#' (see <https://www.smartatlantic.ca/station_alt.html?id=halifax_h1>)
+#'
+#' * `"halifax"` for a buoy in Herring Cove, Nova Scotia
+#' (see <https://www.smartatlantic.ca/station_alt.html?id=halifax>)
+#'
+#' * `"hkb"` for a buoy near Meagher's Beach, Nova Scotia
+#' (see <https://www.smartatlantic.ca/station_alt.html?id=halifax_hk4>)
+#'
+#' * `"saint_john"` for a buoy in the Bay of Fundy, near St John, New Brunswick
+#' (see <https://www.smartatlantic.ca/station_alt.html?id=saintjohn>)
+#'
+#' * `"saint_johns"` for a buoy near St John's Harbour, Newfoundland
+#' (see <https://www.smartatlantic.ca/station_alt.html?id=stjohns>)
 #'
 #' @param program argument specifying the desired oceanographic
-#' program to download buoy data from. If this is not supplied,
-#' `dod.buoy()` will print a list of possibilities.
+#' program to download buoy data from. This must be either
+#' `"MEDS"` or `"smartatlantic"`.
 #'
-#' @param ID a character value indicating the ID of the instrument.
+#' @param ID a character value indicating the ID of the instrument (see \sQuote{Details})
 #'
 #' @template destdirTemplate
 #'
@@ -19,11 +50,11 @@
 #' @importFrom utils unzip
 #'
 #' @return If `index` is TRUE, and `program` is `"BBMP"` or `"BATS"`,
-#' return a data frame.  Otherwise, return the name of the downloaded file.
+#' return a data frame.  Otherwise, return the name of the
+#' downloaded file.
 #'
 #' @examples
-#' # Show recent variation of significant wave height in Halifax
-#' # Harbour.
+#' # Show significant wave height in Halifax Harbour over past 28 days.
 #' # Note that a temporary directory is used, in case
 #' # the package is later submitted to CRAN, which does not
 #' # permit downloads to the working directory.
@@ -34,7 +65,9 @@
 #' col.names <- strsplit(readLines(file, 1), ",")[[1]]
 #' d <- read.csv(file, skip = 2, col.names = col.names)
 #' d$t <- as.POSIXct(d$time, tz = "UTC", format = "%Y-%m-%dT%H:%M:%SZ")
-#' plot(d$t, d$wave_ht_sig, type = "l",
+#' look <- d$t > (max(d$t, na.rm = TRUE) - 28 * 86400)
+#' plot(d$t[look], d$wave_ht_sig[look],
+#'     type = "l", xaxs = "i",
 #'     xlab = "", ylab = "Sig. Wave Ht. [m]"
 #' )
 #' unlink(tmpdir, recursive = TRUE) # remove temporary directory
@@ -44,7 +77,7 @@
 #' @author Dan Kelley
 # https://www.meds-sdmm.dfo-mpo.gc.ca/alphapro/wave/waveshare/csvData/c44258_csv.zip
 # https://www.meds-sdmm.dfo-mpo.gc.ca/alphapro/wave/waveshare/csvData/c44139_csv.zip
-dod.buoy <- function(program, ID = NULL, destdir = ".", age = age, debug = 0) {
+dod.buoy <- function(program, ID = NULL, destdir = ".", age = 1, debug = 0) {
     programAllowed <- c("MEDS", "smartatlantic")
     if (missing(program)) {
         stop("must supply 'program'; try one of: \"", paste(programAllowed, collapse = "\" \""), "\"")
@@ -54,7 +87,6 @@ dod.buoy <- function(program, ID = NULL, destdir = ".", age = age, debug = 0) {
     }
     dodDebug(debug, "dod.buoy(program=\"", program, "\", ID=\"", ID, "\", ...)\n", sep = "")
     if (program == "MEDS") {
-        dodDebug(debug, "handling program \"MEDS\"\n")
         loc <- list(
             "East Scotian Slope" = "44137",
             "Banquereau Bank" = "44139", "Halifax Harbour" = "44258",
@@ -80,45 +112,91 @@ dod.buoy <- function(program, ID = NULL, destdir = ".", age = age, debug = 0) {
         dodDebug(debug, "downloaded \"", rval, "\"\n", sep = "")
         return(rval)
     } else if (program == "smartatlantic") {
-        dodDebug(debug, "handling program \"smartlatlantic\"\n")
-        if (grepl("h1", ID)) {
-            dodDebug(debug, "handling \"h1_buoy\"\n")
-            file <- "XEOS_h1_buoy.csv"
-            url <- paste0(
-                "https://www.smartatlantic.ca/",
-                "erddap/tabledap/SmartAtlantic_",
-                file,
-                "?",
-                "time,",
-                "latitude,",
-                "longitude,",
-                "temperature,",
-                "wave_ht_sig,",
-                "wave_period_max,",
-                "wave_ht_max,",
-                "wave_dir_avg,",
-                "wave_spread_avg,",
-                "sea_surface_wave_mean_period,",
-                "sample_quality"
-                #"&time>=2024-03-22T00:00:00Z",
-                #"&time<=2024-03-29T22:18:00Z"
+        IDabbrev <- list(
+            "h1" = "SmartAtlantic_XEOS_h1_buoy",
+            "hkb" = "SmartAtlantic_XEOS_hkb_buoy",
+            "halifax" = "SMA_halifax",
+            "saint_john" = "SMA_saint_john",
+            "saint_johns" = "SMA_saint_johns"
+        )
+        IDallowed <- names(IDabbrev)
+        if (!ID %in% IDallowed) {
+            stop(
+                "ID \"", ID, "\" not permitted; try one of \"",
+                paste(IDallowed, collapse = "\" \""), "\""
             )
-            filename <- paste0(destdir, "/", file)
-            # The next two lines showed that the curl:: method is 20%
-            # faster in elapsed time, although 2.5X faster in user time;
-            # Frankly, either would be fine, because elapsed time is
-            # what the user sees. I will stick with download.file()
-            # because it does not force a dependence on another
-            # library.
-            #<SPEED TEST> print(system.time(curl::curl_download(url, filename)))
-            #<SPEED TEST> print(system.time(download.file(url, filename)))
-            #curl::curl_download(url, filename)
-            download.file(url, filename)
-            dodDebug(debug, "downloaded \"", filename, "\"\n", sep = "")
-            return(filename)
-        } else {
-            stop("ID=\"", ID, "\" not recognized for program=\"smartatlantic\"; try \"h1\"")
         }
+        # Access file directly (will this always work, though?)
+        base <- "https://www.smartatlantic.ca/erddap/files/"
+        if (ID == "h1") {
+            #https://www.smartatlantic.ca/erddap/files/SmartAtlantic_XEOS_h1_buoy/smb_h1_data.csv
+            #dodDebug(debug, "handling \"", ID, "\" as a direct download\n", sep = "")
+            file <- "smb_h1_data.csv"
+            url <- paste0(base, "SmartAtlantic_XEOS_h1_buoy/", file)
+            return(dod.download(url = url, file = file, age = age, destdir = destdir, debug = debug - 1))
+        }
+        if (ID == "halifax") {
+            #https://www.smartatlantic.ca/erddap/files/SMA_halifax/smb_halifax.csv
+            #dodDebug(debug, "handling \"", ID, "\" as a direct download\n", sep = "")
+            file <- "smb_halifax.csv"
+            url <- paste0(base, "SMA_halifax/", file)
+            return(dod.download(url = url, file = file, age = age, destdir = destdir, debug = debug - 1))
+        }
+        if (ID == "hkb") {
+            # https://www.smartatlantic.ca/erddap/files/SmartAtlantic_XEOS_hkb_buoy/smb_hkb_data.csv
+            #dodDebug(debug, "handling \"", ID, "\" as a direct download\n", sep = "")
+            file <- "smb_hkb_data.csv"
+            url <- paste0(base, "SmartAtlantic_XEOS_hkb_buoy/", file)
+            return(dod.download(url = url, file = file, age = age, destdir = destdir, debug = debug - 1))
+        }
+        if (ID == "saint_john") {
+            #dodDebug(debug, "handling \"", ID, "\" as a direct download\n", sep = "")
+            file <- "smb_saint_john.csv"
+            url <- paste0(base, "SMA_saint_john/", file)
+            return(dod.download(url = url, file = file, age = age, destdir = destdir, debug = debug - 1))
+        }
+        if (ID == "saint_johns") {
+            #https://www.smartatlantic.ca/erddap/files/SMA_st_johns/smb_st_johns.csv
+            #dodDebug(debug, "handling \"", ID, "\" as a direct download\n", sep = "")
+            file <- "smb_st_johns.csv"
+            url <- paste0(base, "SMA_st_johns/", file)
+            return(dod.download(url = url, file = file, age = age, destdir = destdir, debug = debug - 1))
+        }
+        stop("Unknown ID \"", ID, "\" (programming error: please reort this)")
+        #<old> IDfull <- IDabbrev[ID]
+        #<old> dodDebug(debug, "handling \"", ID, "\", i.e. \"", IDfull, "\"\n")
+        #<old> file <- paste0(IDfull, ".csv")
+        #<old> url <- paste0(
+        #<old>     "https://www.smartatlantic.ca/erddap/tabledap/",
+        #<old>     file,
+        #<old>     "?",
+        #<old>     "time,",
+        #<old>     "latitude,",
+        #<old>     "longitude,",
+        #<old>     "temperature,",
+        #<old>     "wave_ht_sig,",
+        #<old>     "wave_period_max,",
+        #<old>     "wave_ht_max,",
+        #<old>     "wave_dir_avg,",
+        #<old>     "wave_spread_avg,",
+        #<old>     "sea_surface_wave_mean_period,",
+        #<old>     "sample_quality"
+        #<old>     # "&time>=2024-03-22T00:00:00Z",
+        #<old>     # "&time<=2024-03-29T22:18:00Z"
+        #<old> )
+        #<old> filename <- paste0(destdir, "/", file)
+        #<old> # The next two lines showed that the curl:: method is 20%
+        #<old> # faster in elapsed time, although 2.5X faster in user time;
+        #<old> # Frankly, either would be fine, because elapsed time is
+        #<old> # what the user sees. I will stick with download.file()
+        #<old> # because it does not force a dependence on another
+        #<old> # library.
+        #<old> #<SPEED TEST> print(system.time(curl::curl_download(url, filename)))
+        #<old> #<SPEED TEST> print(system.time(download.file(url, filename)))
+        #<old> # curl::curl_download(url, filename)
+        #<old> download.file(url, filename)
+        #<old> dodDebug(debug, "downloaded \"", filename, "\"\n", sep = "")
+        #<old> return(filename)
     } else {
         stop("unrecognized program=\"", program, "\"; try one of: \"", paste(programAllowed, collapse = "\" \""), "\"")
     }