+++ sandbox/dek/buoy/test_buoy.R
@@ -1,0 +1,5 @@
+library(dod)
+for (loc in c("h1", "halifax", "hkb", "saint_john", "saint_johns")) {
+    cat("location: \"", loc, "\"\n", sep = "")
+    dod.buoy("smartatlantic", loc, debug = 2)
+}